pyproject.toml
[project]
name = "generative-search-engine-for-local-files"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "faiss-cpu>=1.9.0",
    "numpy>=2.1.2",
    "ollama>=0.3.3",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "python-pptx>=1.0.2",
    "requests>=2.32.3",
    "sentence-transformers>=3.2.0",
    "streamlit-lottie>=0.0.5",
    "streamlit>=1.39.0",
]
